gui modules: rename icon_menu -> menu
