+++ applications/gui/modules/menu.c
@@ -1,4 +1,4 @@
-#include "icon_menu.h"
+#include "menu.h"
 
 #include <m-array.h>
 #include <gui/elements.h>
